refactor: use in memory implementation only
